test(verify): the corpus interpreters are shared test support
The two interpreters of a golden — the bus-driven engine and the hand
driver of AgentRun — and the Program table they replay move from
golden_replay.rs into tests/corpus/mod.rs, where every matrix module
includes them. The table gains the axes the matrices vary: an absent or
appended preamble, static context, a prior history, max_tokens,
additional_params, tool_choice, an output schema, and hooks by name.
The module doc holds the dimension table of an effect trace and the
census of what the original ten goldens cover.
